add stock prices endpoint

+++ execution-service/src/main/java/execution_service/requests/InternalDeleteStockTXRequest.java
@@ -2,4 +2,4 @@ package execution_service.requests;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record InternalDeleteStockTXRequest(@JsonProperty("stock_tx_id") Integer stockTXId) {}
+public record InternalDeleteStockTXRequest(@JsonProperty("stock_tx_id") Long stockTXId) {}
